refactor(major):  major refactor on all the logics of pr-generator and analysis
