+++ src/git-utils.ts
@@ -8,12 +8,6 @@ import type {
 
 const git = simpleGit();
 
-/**
- * Get the git changes between the current branch and a specified base branch.
- * @param baseBranch Base branch to compare against
- * @param maxFiles Maximum number of files to include in the changes
- * @returns A promise that resolves to an object containing git changes information.
- */
 export async function getGitChanges(
   baseBranch: string,
   maxFiles: number
