+++ src/utils.ts
@@ -1,10 +1,5 @@
-/**
- * Masks the middle part of an API key, showing only the first and last 4 characters.
- * @param apiKey The API key to mask.
- * @param visibleChars The number of characters to show at the start and end of the API key.
- *
- * @returns The masked API key.
- */
+import { DetectedPattern, PatterTypes } from "./types";
+
 export function maskApiKey(apiKey: string, visibleChars = 4): string {
   if (!apiKey || apiKey.length <= visibleChars * 2) {
     return apiKey; // No masking needed if the key is too short
@@ -17,13 +12,6 @@ export function maskApiKey(apiKey: string, visibleChars = 4): string {
   return `${startVisible}${maskedPart}${endVisible}`;
 }
 
-/**
- * Validate file patterns against file path
- * @param filePath The file path to check against the patterns.
- * @param patterns An array of glob patterns to match against the file path.
- *
- * Returns true if the file path matches any of the patterns, false otherwise.
- */
 export function matchesFilePattern(
   filePath: string,
   patterns: string[]
@@ -42,12 +30,6 @@ export function matchesFilePattern(
   });
 }
 
-/**
- * Format file size in human readable format
- * @param bytes The size in bytes to format.
- *
- * Returns a string representing the size in B, KB, MB, or GB.
- */
 export function formatFileSize(bytes: number): string {
   const sizes = ["B", "KB", "MB", "GB"];
   if (bytes === 0) return "0 B";
@@ -56,25 +38,11 @@ export function formatFileSize(bytes: number): string {
   return `${Math.round((bytes / Math.pow(1024, i)) * 100) / 100} ${sizes[i]}`;
 }
 
-/**
- * Truncate text to specified length
- * @param text The text to truncate.
- * @param maxLength The maximum length of the text.
- *
- * Returns the truncated text, appending "..." if it exceeds maxLength.
- */
 export function truncateText(text: string, maxLength: number): string {
   if (text.length <= maxLength) return text;
   return text.substring(0, maxLength - 3) + "...";
 }
 
-/**
- * Get relative time string
- * @param date The date to compare against the current time.
- *
- * Returns a string representing the time difference in a human-readable format.
- * For example, "just now", "5m ago", "2h ago"
- */
 export function getRelativeTime(date: string | Date): string {
   const now = new Date();
   const target = new Date(date);
@@ -91,3 +59,128 @@ export function getRelativeTime(date: string | Date): string {
 
   return target.toLocaleDateString();
 }
+
+export const reviewPatters = {
+  // Security vulnerability patterns
+  securityPatterns: [
+    /eval\s*\(/gi, // eval() usage
+    /innerHTML\s*=/gi, // innerHTML assignment (XSS risk)
+    /document\.write\s*\(/gi, // document.write (XSS risk)
+    /\$\{[^}]*\}/g, // Template literal injection risk
+    /SELECT\s+.*\s+FROM\s+.*\s+WHERE\s+.*\+/gi, // SQL injection pattern
+    /password\s*=\s*['"]/gi, // Hardcoded passwords
+    /api[_-]?key\s*=\s*['"]/gi, // Hardcoded API keys
+    /secret\s*=\s*['"]/gi, // Hardcoded secrets
+    /token\s*=\s*['"]/gi, // Hardcoded tokens
+    /crypto\.createHash\s*\(\s*['"]md5['"]/gi, // Weak hashing
+    /Math\.random\s*$$\s*$$/gi, // Insecure random for security
+  ],
+  // Performance issue patterns
+  performancePatterns: [
+    /for\s*$$[^)]*$$\s*{[^}]*for\s*\(/gi, // Nested loops
+    /while\s*$$[^)]*$$\s*{[^}]*while\s*\(/gi, // Nested while loops
+    /\.map\s*$$[^)]*$$\.map\s*\(/gi, // Chained array operations
+    /JSON\.parse\s*\(\s*JSON\.stringify/gi, // Inefficient deep clone
+    /new\s+RegExp\s*\(/gi, // RegExp in loops (should be outside)
+    /console\.log\s*\(/gi, // Console logs in production
+    /debugger\s*;/gi, // Debugger statements
+    /alert\s*\(/gi, // Alert statements
+  ],
+
+  // Bug-prone patterns
+  bugPatterns: [
+    /==\s*null/gi, // Should use === null
+    /!=\s*null/gi, // Should use !== null
+    /==\s*undefined/gi, // Should use === undefined
+    /!=\s*undefined/gi, // Should use !== undefined
+    /\+\+\w+\[/gi, // Increment with array access
+    /\w+\[\+\+/gi, // Array access with increment
+    /catch\s*$$[^)]*$$\s*{\s*}/gi, // Empty catch blocks
+    /if\s*$$[^)]*$$\s*;\s*$/gm, // Empty if statements
+    /else\s*;\s*$/gm, // Empty else statements
+  ],
+
+  // Style and maintainability patterns
+  stylePatterns: [
+    /function\s+\w+\s*$$[^)]*$$\s*{[\s\S]{500,}?}/gi, // Large functions
+    /class\s+\w+\s*{[\s\S]{1000,}?}/gi, // Large classes
+    /\/\*[\s\S]*?\*\//g, // Block comments (check for TODO/FIXME)
+    /\/\/\s*(TODO|FIXME|HACK|XXX)/gi, // Technical debt comments
+    /var\s+/gi, // var usage (should use let/const)
+  ],
+};
+
+export const patternsInfo: Record<
+  Exclude<PatterTypes, "maintainability">,
+  Record<
+    string,
+    {
+      severity: DetectedPattern["severity"];
+      description: string;
+      suggestion: string;
+    }
+  >
+> = {
+  security: {
+    "eval\\s*\\(": {
+      severity: "critical",
+      description: "Use of eval() can lead to code injection vulnerabilities",
+      suggestion:
+        "Avoid eval() and use safer alternatives like JSON.parse() for data",
+    },
+    "innerHTML\\s*=": {
+      severity: "high",
+      description:
+        "Direct innerHTML assignment can lead to XSS vulnerabilities",
+      suggestion: "Use textContent or sanitize HTML content before assignment",
+    },
+    "password\\s*=\\s*['\"]": {
+      severity: "critical",
+      description: "Hardcoded password detected in source code",
+      suggestion:
+        "Use environment variables or secure configuration for passwords",
+    },
+  },
+
+  bug: {
+    "==\\s*null": {
+      severity: "medium",
+      description: "Loose equality with null can cause unexpected behavior",
+      suggestion: "Use strict equality (===) for null checks",
+    },
+    "catch\\s*$$[^)]*$$\\s*{\\s*}": {
+      severity: "high",
+      description:
+        "Empty catch blocks hide errors and make debugging difficult",
+      suggestion: "Add proper error handling or at least log the error",
+    },
+  },
+  performance: {
+    "for\\s*$$[^)]*$$\\s*{[^}]*for\\s*\\(": {
+      severity: "medium",
+      description:
+        "Nested loops can cause performance issues with large datasets",
+      suggestion:
+        "Consider optimizing algorithm or using more efficient data structures",
+    },
+    "console\\.log\\s*\\(": {
+      severity: "low",
+      description: "Console logs should be removed from production code",
+      suggestion:
+        "Remove console.log statements or use proper logging framework",
+    },
+  },
+
+  style: {
+    "var\\s+": {
+      severity: "low",
+      description: "var has function scope and can cause confusion",
+      suggestion: "Use let or const for block-scoped variables",
+    },
+    "\\/\\/\\s*(TODO|FIXME|HACK|XXX)": {
+      severity: "low",
+      description: "Technical debt comment indicates incomplete work",
+      suggestion: "Address the TODO/FIXME or create a proper issue to track it",
+    },
+  },
+};
